bug修正
